+++ src/interfaces/ValidatorFn.ts
@@ -1,0 +1,6 @@
+import { TypeValue } from "../decorators/types";
+
+export type ValidatorFn<T extends object> = (
+  argValue: T | undefined,
+  argType: TypeValue,
+) => void | Promise<void>;