+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 <!-- here goes all the unreleased changes descriptions -->
+### Features
+- allow passing custom validation function as `validate` option to `buildSchema`
 ### Fixes
 - don't include in schema the fields declared as `@FieldResolver` when that resolvers classes aren't provided in `resolvers` array